requirements.txt
annotated-types==0.7.0
certifi==2026.1.4
charset-normalizer==3.4.4
fastapi==0.116.1
google-api-python-client==2.182.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
idna==3.11
inflection==0.5.1
jinja2==3.1.6
pyairtable==3.3.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
python-multipart==0.0.20
requests==2.32.5
typing-inspection==0.4.2
typing_extensions==4.15.0
numpy==2.4.2
opencv-python==4.13.0.92
boto3>=1.35.0
urllib3==2.6.3
uvicorn==0.35.0
youtube-transcript-api==1.2.4
